pyproject.toml
[tool.poetry]
name = "animecal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
icalendar = "^4.0.7"
selectolax = "^0.3.1"
requests = "^2.26.0"
Flask = "^2.0.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
